Changes made to work around bcc compiler errors:
Removed 'using' statements for partially overloaded value() methods in typed_array_field and typed_scalar_field subclasses. I don't believe they are needed, as its OK for the other overloads to be hidden in the subclass.

Also renamed one of the read and write methods in field_formatter to avoid method overloading.