Remove VXL_SOURCE macro from vxl_version.h
This macro was defined to a string literal containing either "RELEASE"
or "CVS".  Since we switched to Subversion the name "CVS" makes no
sense.  Instead of switching to "SVN" or something we've decided to
remove the macro completely.  No preprocessor tests can be conducted
against the macro, and the other version macros fully distinguish the
version anyway.  Fixes #11.

SVN-Revision: 24541