Added ATOMICS_FLOAT functionality to boxm2 update process.
This allows processing large numbers of rays per voxel, which occurs when there is a discrepancy between voxel and pixel resolutions. In such cases the old method may result in overflows causing the update to fail. The new method can be enabled by adding the flag -D ATOMICS_FLOAT to the ocl compilation options.