made metadata parsing stricter
n_bands_, gain_, offset_ (or gains_) are assumed to be parsed from the metadata file or, in the case of IKONOS, hardcoded, although this is only enforced in parse_from_txt() at the moment. if solar_irradiance cannot be set, then this method fails now.
fixed bugs when processing an IKONOS image: neither n_bands_ nor solar irradiance was being set.
added SunAngle as a tag in parse_from_txt()
added a default value of -1 to gsd_