avoid confusion compiler errors in VS2013
When compiling with Visual Studio 2013 Professional, if bstm_ocl_render_view is configured to compile (opencl, glew and vgui are found) then I get confusing compiler errors about undefined type, int, during a macro expansion in vnl_math. It seems like something is clobbering the namespace or something. This isn't a problem in boxm2_ocl_render_view, from which this exe was copied.