added a pyadaptor to setup dbvalue as a smart pointer
boxm2_smart_register sets up boxm2_register.dbvalue as a smart_pointer, capable of removing its reference from the C++-side database when all references to the object fall out-of-scope. if used, this setup routine should be called before other adaptors. it works by sideloading a __del__() method on dbvalue and makes the boxm2_register.remove_data() function a no-op; boxm2_batch.remove_data() should not be explicitly called with a dbvalue.id as this will lead to a double free.