Added VOLM_DIR env var
Now boxm2_batch (and anything including volm) is stable when moving the code/binaries. You can set a VOLM_DIR env var to identify the new location of the txt files in the volm src dir needed at run time. In addition, if the files are missing, it no longer enters an infinite loop.