Refactor and fix istream frame numbering
The previous frame numbering in istream was based on the DTS (decode
timestamp). It should really be based off the PTS (presentation timestamp).
Note that the start_time_ is also the PTS start time of the video and not
related to the DTS.
Most of the changes are refactoring to use FFMPEG function to deal with
quotient better. These handle overflows so we don't have to worry about it.
The variable start_time was renamed to start_time_ for coherence with all
the other pimpl variables.