Enable the use of CMAKE_WINDOWS_EXPORT_ALL_SYMBOLS to build shared with MSVC.
This commit allows VXL to use the new CMake feature which exports symbols
from all object files when creating a DLL.  With this change only data
exported from the dll needs to have mark up.  This commit addresses the
exported data in VXL. In addition, several places where libraries were not
fully listing dependent libraries were fixed. There is still an issue with
mbl_cloneables_factory as it uses a static object list in a templated class.
This can cause duplicate copies of that static object across dll's and exe's.
That breaks the singleton pattern and the code.  mcal_test_pca,
mcal_test_trivial_ca, mcal_test_general_ca, and mfpf_test_edge_finder fail
because of this. vidl_test_color fails because it checks on a pointer to
a function which gets duplicated accross the dll and exe, but the code
works.  All tests that passed before this change still pass in a static build
after this change.