bocl_manager now returns reference to singleton instance
As part of this change, bocl_manager smart pointers are no longer
created, and can no longer be stored in the brdb database.  Instead,
users should call the singleton instance() method to get a reference
to the bocl_manager<T> instances.

This fixes the crashes happening when using the boxm2_smart_register
smart pointers in python scripts, and is also safer in general.