removed boxm2_cache_destroyer class
This commit fixes a crash on exit caused by the boxm2_cache_destroyer
and static instance_ smart_ptr both attempting to delete the
boxm2_cache object pointed to by instance_.

The purpose of the boxm2_cache_destroyer class was to make sure the
singleton boxm2_cache instance was deleted on program exit.
At some point, however, the static instance_ pointer was changed to
a vbl_smart_ptr, which will be destructed on program exit, causing
ref_count to drop to 0, and the boxm2_cache instance to be deleted
without help from the destroyer.