Added ability for stream cache to continue when files missing
Instead of having a stream cache thrown an exception when a file is
missing from any appearance model, as long as at least one exists,
the output vector will contain only the values for appearance models
that were successfully loaded. Previously an exception 0 would have
been thrown, but now it is actually possible that a vector of smaller
size is returned. So be careful if that matters. Also, if NON
appearance models are found for a specific block, the logic become
VERY inefficient and tries to open every appearance model for every
cell. To prevent that, exists was added to boxm2_stream_cache so that
the main block loop can determine if there are enough appearance
models to continue. 0 should always be skipped. Depending on your algo
other scenarios should be skipped. But it's better then crashing.