COMP: Fix VNL compilation for wider C++11 support
Initial code added to support C++11 worked with GCC 4.8, but with GCC 4.9
a more strict enforcement of the requirements is enforced.

This was previous reported in https://issues.itk.org/jira/browse/ITK-2976
ITK/Modules/ThirdParty/VNL/src/vxl/core/vnl/vnl_numeric_traits.h:387:27:
error: ‘constexpr’ needed for in-class initialization of static data member ‘zero’ of non-integral type

When building for C++11 and optionally using constexpr instead of const
it is important that the vxl_platform_tests.cxx reflect a consistent
view of the compiler features.

A more complete C++11 compliant solution is presented in this patch set.