STYLE: REMOVE EGCS compiler
The EGCS compiler is no longer supported.
This patch removes the code specialization required
to use VXL with that compiler.

In April 1999, after a long period of behind-the-scenes negotiation,
EGCS and GCC reunited, and EGCS was accepted as the GNU Project's
official GCC. At that time, GCC was renamed from "GNU C Compiler" to
"GNU Compiler Collection", which conveniently has the same acronym. The
first compiler version released after the reunification was GCC version
2.95.

NOTE: GCC 2.95 is will also be removed soon, as it is also not supported
any longer.