COMP:  putback is not reliable for istreams
The istream interface explicitly states[1] the putback (& unget, and
all other methods that write to the stream) :
"... may either fail, be ignored, or overwrite the character at that position"
                ^^^^^^^^^^^^^^^^^

In the case of the clang compiler, it was failing.

A new implementation was created that does not require writing to the
stream.  This new implementation passes all the tests.

[1] http://www.cplusplus.com/reference/istream/istream/putback/