Define OF for minizip, when it isn't already.
Newer zlib has changed its OF macro to _Z_OF. This branch first checks if OF
is not defined. If not, check for _Z_OF and use that definition if
available. Otherwise, use default definition.

A better patch would update the minizip code in brl, but I don't know if
that's too radical an approach.