test 3D point instead of (non-unique) 2D point
This test sometimes failed, even though the returned line was
technically correct.  The problem was that the "uv" basis vectors are
arbitrary when the direction is very close to the x axis, meaning the
x0_ point being tested is also arbirary.  The 3D point closest to the
origin is uniquely defined, so test that instead.  In the future we
should probably remove x0_ from the public interface for the same
reason.