STYLE: Remove vnl_brent_minimizer::minimize_golden
This function is not used anywhere in vxl (or ITK).

The function documentation makes a claim that fc is used,
but static analysis identified this variable as read but not
used:
NOTE:  warning: Value stored to 'fc' is never read [clang-analyzer-deadcode.DeadStores]

Inspection of the code indicates that this function does not perform
as the documentation states.  Further investigation indicates that this
function is never used in vxl (or ITK).