STYLE: Simply by removing unnecessary sentinal value
vxl/core/vnl/algo/vnl_sparse_lm.cxx:156:7: note: Value stored to 'stop' is never read
      stop = true;
      ^

The intent is to stop iterating throught the for loop when these conditions are met.

The stop variable could be used if it was followed by "continue", but since it is followed
by "break", the for loop never checks the value when it is set to true.