COMP: BRL code is unstable on different compilers
The BRL code depends on non-standard c++ compiler behaviors
such as zero initializing variables implicitly.  This is not
the case for clang (or other compilers) and causes many tests to
either fail, segfault, or enter infinite loops and time out.

https://github.com/vxl/vxl/issues/54

A few of these issues were identified and corrected with
static analysis, but the scope of the issues is too large
to fix quickly.

This patch disables BRL by default until a more stable
infrastructure can be put into place.  BRL will still be built
by default when VXL_BUILD_DASHBOARDS is turned on or if USE_BRL
is turned on explicitly.