ENH: The iso646 headers are not really needed.
In vxl these end up mapping to empty header files.

In C++, reserved words exist with the same names as these macros and are treated as aliases of their respective operator. Therefore the inclusion of this header has no effect in C++, and is not necessary in order to use the alternative names.