STYLE: Remove SunPro support
The conditional compilation logic for SUNPRO_CC5[056] versions
(5.0, 5.5, 5.6 ) greatly convoluted the code.  The newest versions
of the SunPro comiplers (5.13 aka 12.4) definitely conflict with
the old work arounds.  Removing the old conditional compilation
work arounds provided improved compatibility with the new
5.13 compiler on linux.

There were several new challenges with the the 5.13 compiler that
were not addressed in the 5.5 and 5.6 workarounds.

===
SUNPRO:  Default build uses libcstd++ does not support iterator_traits
http://www.oracle.com/technetwork/server-storage/solaris/cmp-stlport-libcstd-142559.html

Can build with stlport4, but that cautioned against in the documentation

"vxl/core/vnl/vnl_matrix.h", line 54: Warning: Export ignored -- not yet implemented.
===

Current versions of vxl will not build against any SunPro version, so
this patch is removing the outdated conditional compilation code. If
support for the 5.13 compiler is needed in the future, it will be better
to address those issues independantly.