ENH: Provide heirarchial build options
Using the cmake 2.6 feature of CMAKE_DEPENDENT_OPTION the
build choices and configuration of which modules to build
is made more robust and easier to navigate.

Previously, options configurations would allow one to choose
options that had no effect, and it was therefore unclear
which combinations of options were necessary in order to
achieve a particular built solution.