COMP: Broken test when BUILD_CORE_UTILITIES=OFF
When -DBUILD_CORE_UTILITIES:BOOL=OFF was used, a few
tests with inter-module dependancies were failing to compile.

A new test driver was added that is conditionally used based
on CMake choices.

ENH: Timing test was missing for timing test application.

The following configuration now builds the part of VXL
that used by ITK community and tests that part independant
of all other contributions.

cmake \
  -DBUILD_CORE_GEOMETRY:BOOL=OFF \
  -DBUILD_CORE_IMAGING:BOOL=OFF \
  -DBUILD_CORE_SERIALISATION:BOOL=OFF \
  -DBUILD_CORE_UTILITIES:BOOL=OFF \
  -DBUILD_CORE_VIDEO:BOOL=OFF \
  -DBUILD_EXAMPLES:BOOL=OFF \
  -DCMAKE_BUILD_TYPE:STRING=Debug

This will assist with testing VXL's core parts to
allow developers to focus on removing compiler errors
and improving code coverage of this most important
part of the toolkit.