STYLE: Remove custom gcc-libstdcxx-v3
Since gcc version 3, gcc-libstdcxx-v3 has been included
in every compiler.  This was a forward port for the 2.96
gcc compiler. Now that gcc 2.96 is not supported, we can
simplify the code and remove the alias and conditionals
related to these specializations.

Templates/fstream-instances.cxx is no longer referenced
or used.

The following conditional could no longer be true:
\#if defined(VCL_GCC) && !defined(GNU_LIBSTDCXX_V3)

Specialized functions template definitions for when
GCC version was less than 4 are eliminiated.