ENH: Adds support for automatic configuration of Coverage and Memory Checking
Configuring a build to submit coverage data requires specific compiler
and linker flags to be set. Previously setting the dashboard_do_coverage
variable only caused the ctest script to try to submit coverage, even if
the project was not built with the correct settings. Now, when asking the
ctest script to do coverage, compiler and linker flags are set, and a coverage
command is searched for as well. If no coverage command is found the test
script will fail. This makes it easier to setup a coverage enabled dashboard
submission.
If dashboard_do_memcheck is set, valgind is automatically found,
and the valgind suppressions file is automatically loaded. Previously
this would have had to have been configured manually.