BUG: Value required for variable used in header define
check_cxx_expression_compiles macro was returning "" instead of 0
for failure case.

This was an artifact of CHECK_CXX_SOURCE_COMPILES return 1 on success,
"" on failure.

The resultant macro definitions in vxl_config.h becomes un-valued, as
shown below.  This leads to compile error c1017
(https://msdn.microsoft.com/en-us/library/h5sh3k99.aspx) in VS11.

The solution was to ensure that the variable had a value of 0
on failure.