COMP: Fix external installs of v3p/clipper
The v3p/clipper package was the only version
that used the .hpp extension. Rather than
extend the pattern matching regex for installing
files, this file name was changed to use
the .hxx extension.

The clipper package was also updated to the
latest version as documented in the new
UpstreamNotes.txt file.