STYLE: Remove deprecated options to simplify debugging
Internal comment related to this code added in 2003 is:
"These should be removed after some time to upgrade has passed."

VXL_PROVIDE_OLD_CACHE_NAME was added in 2003 to provide
a transition period for external projects. The 13 year
transition period is now over, and deprecated variables
are being removed.

Many of these deprecated variables are not similar
to existing variables, or no longer make sense to
have with cmake 2.8.12 conventions.