ENH: When possible, prefere C++11 functions
C++11 now provides functions that had previously required
specialization.  When compiling with C++11 or greater, prefer
to use the std:: version of the function to the locally provided
versions.

hypot, cbrt, min, max, isnan, isinf, isfinite, isnormal are
currently prefered to be provided by C++11 when available.

The C++ functions are called using perfect forwaring so that no
performance penalty is encountered.

NOTE:  C98 emulation of these functions is retained for backwards
compatibility.