ENH: Simplify travis build requirements
VXL is capable of being compiled with the default
cmake 2.8.7 version that is available by default
in the travis CI envirionment. Removing need for cmake 2.8.9
to improve backwards compatibility.

Remove custom install of CMake 2.8.9 and the need for a
cache build.

Add ability to report to dashboard each build. We already
configure, build, and test for each platform, so submitting
those results to the nightly experimental dashboards is
not much of an extra burdon.