ENH: Adds CMake macro combining  PERFORM_CMAKE_TEST and SET_INVERT
The two macros, PERFORM_CMAKE_TEST and SET_INVERT are often used
together when a false result is expected. This macro combines both
of these together. NOTE: this does not replace the exact function of
the two old macros with the new one. Because PERFORM_CMAKE_TEST skips
its test if the TEST variable is defined, calling the two functions
in succession might have resulted in flipping the desired result, as
SET_INVERT did not know that PERFORM_CMAKE_TEST had decided not to
test. This was assumed to be a bug, and is addressed in the new macro
that combines the two functions.