ENH: Simplify std:: math function definitions
In the early to mid 2000's C++ compilers had
many variations and shortcommings.  A set of
vcl_* functions were made to provide a consistent
interface to the C++98 standard functions and
features.  Currently all VXL supported compilers
are now compliant and consistently provide
interfaces to these functions, so there
is no need for the complicated include files
or the complicated optional code paths.

This patch set identified the only valid
code path and greatly simplifies the conditional
logic so that maintenance is easier.

Additionally, it demonstrates that packages
that depend on this version of VXL could simply
use the C++ provided version of these functions
with out using the aliases.