ENH: Prefer std:: over vcl versions.
The vcl_* definitions can now be greatly simplified.
After removing specializations for early non-conformant
c++ compilers the end result was that only the
std:: version of the functions could ever be
used by the compiler.

Continued review indicated that this could be greatly simplified
to be simple aliases to the std:: versions.