COMP: Supress jpeg deprecation warning
There is no need to report this warning. It just makes the dashboard noisy.
If we explcitly set USE_OPJ_DEPRECATED to on, then we have determined that we
want to use the deprecated function, and we should not warn about it.

A complete update of JPEG implementation is required and documented in
https://github.com/vxl/vxl/issues/229