STYLE: Step toward external backwards compatiblity.
In order to provide external backwards compatibility
across applications that depended on transitive
header include behavior, the #define aliases for
vcl_* functions need to be visible from all previous header
files.

The approach here is to consolodate those aliases in one place
so that it is easier to disable those aliases for external application
developers that wish to modernize thier code.