BUG: Fix VXL header INSTALL_INTERFACE directory.
This addresses two issues. First, for ITK, which has a non-standard
VXL_INSTALL_INCLUDE_DIR, the headers are install to the correct directory
(${VXL_INSTALL_INCLUDE_DIR}).  This addresses:

CMake Error in CMakeLists.txt:
  Imported target "itkvnl_algo" includes non-existent path

    "/usr/local/include/vxl/core/vnl/algo"

  in its INTERFACE_INCLUDE_DIRECTORIES.  Possible reasons include:

  * The path was deleted, renamed, or moved to another location.

  * An install or uninstall procedure did not complete successfully.

  * The installation package was faulty and references files it does not
  provide.

when building against an installation tree.

Second, CMAKE_PREFIX_PATH should not be used in the target_include_directories
INSTALL_INTERFACE reference because it breaks the use of DESTDIR, etc. -- this
path should be a relative path.