ENH: Provided static API for external applications
When an external application builds against VXL, the API used
should be identical as when VXL was built.  Conditional code paths (#if/#endif)
should be conditioned on static values in the vcl_config_compiler.h file, and *NOT*
solely on compiler internal variables (__cplusplus). This patch identifies
compiler information at configuration time, and then stores that information
for reporting to external users of VXL.