BUG: CMake 2.8.7 bugs found
CMake 2.8.7 bug with CMAKE_SIZEOF_VOID_P

1) A bug in old cmake version may not set CMAKE_SIZEOF_VOID_P so hard
coding for 64 bit OS if it is not properly set. This affects the
ability to detect if the OS is 32 bit or 64 bit.

CMake 2.8.7 bug with GenerateHeaderExport.cmake.  Pathing is
not respected for absolute paths, and static libraries are
not supported with 2.8.7 and hidden visibilty settings. 2.8.9
provides benefits that solutions from ITK and VTK can also be
used in VXL without backporting and supporting extensive cmake
hacks.