ENH: Encourage unified container interface.
VNL containers are not related by inheritance; as such, maintaining
a common interface (where applicable) is not enforced.  Therefore,
container classes may diverge, causing methods to be absent which
the user would logically expect to have.  This patch is a very early
attempt at uncovering and implementing such missing methods through
test-driven-development.  In particular, the following four methods
have been added to vnl_matrix_fixed:

- `swap`
- `is_equal`
- `get_rows`
- `get_columns`

In addition to providing a (more) unified interface, this technique
may reduce the burden in writing unit tests, which currently are
written separately for each container, even when the behavior is
expected to be identical.  The `container_interface` unit test
which is added here provides runtime unit tests only for the
methods which were added (this will hopefully change in the future).