COMP: Remove conditional code for unsupported compilers
On some compilers (in particular gcc 2.7.2.3), the compiler doesn't
know how to cast a templated derived class to a templated base class
(eg. vnl_matrix_fixed<3,3,double> -> vnl_matrix<double>) when doing overload
resolution. Making the overloaded function a friend of the class makes
the problem go away.

This class of compilers is no longer supported, so simplifying the code.