COMP: Remove conditional compilation code
Various outdated compilers do various silly things about instantiation of
functions/methods that have been specialized.  Use this macro to tell
the compiler not to generate code for methods which have been specialized

These compilers are not longer supported, and there is no longer a need or
use of these mechanisms.