COMP: Remove conditional code from unsupported compilers
In order to *define* a template (function or class) specialization, the
definition must be preceded by "template <>" on ISO-conforming compilers.
Some compilers (eg gcc 2.7.2) make no distinction between an instance
of a templated function and a function with the same name and signature,
and so do not support the use of "template <>". Use template <>
instead.

These old non ISO-conforming compilers are not supported any longer.