COMP: Fix VNL tests when cross-compiling
The test name should not have the $<TARGET_FILE:TestCommand> generator
expression. CMake already uses the full path the to the command. This
interferes with the application of CMAKE_CROSSCOMPILING_EMULATOR, which is
required to run the tests successfully.

Also do not run a test that requires file input with Emscripten because it
cannot files on the filesystem by default.