Add standardized error reporting when no ffmpeg is configured.
The vidl_ffmpeg_istream and vidl_ffmpeg_ostream stubs that are used when
no ffmpeg is configured did not always give an indication that the reason
for failure is no ffmpeg rather than some internal reason.

Added error messages and exceptions to make it clear that the caller
is trying to access a stream and there is no ffmpeg support available.