BUG: Convert define 0/1 paradigm to ifdef
The vil_config.h file was configured with a define for VXL_HAS_SSE2_HARDWARE_SUPPORT
that was either 0 or 1. All uses of this symbol were of the #ifdef style.
The result is that the symbol is always defined. This change uses the
#cmakedefine feature which will define the symbol if it has a TRUE
value in cmake. Otherwise it will not be defined. This approach matches
the usage in the code.