COMP: Provide cmake support for C++11
At least CMake 3.8.2 is required so that try_compiles are aware of the
CMAKE_CXX_STANDARD variable.  Without at least this version, compiler
introspection is not consistent across builds. --Brad Lowekamp

This targets the version available with the Ubuntu Long-Term-Support
distribution at the time of the release (this will be 18.04 - Bionic
Beaver). This version is currently CMake 3.9.5. It may change in the
future. Since Windows developers download the latest version from
cmake.org and macOS developers download from cmake.org or use Homebrew
(which tracks latest CMake), there is not a reason to use a lower
version in these cases. --Matt McCormic

VTK, ITK, Slicer have, or are moving to support C++11. Over the recent
history, the maintenance of VXL toolkit has been performed by those
developers.  This change will ensure that integration of VXL
with these packages is less of a burden so that longer term support
will continue to keep VXL working with modern development environments.