Fix vidl visibility
Add missing export decoration so that vidl is usable when hidden
visibility is enabled.

As part of this, fiddle with VIL_SMART_PTR_INSTANTIATE so that the
instantiations can be export-decorated... on not-Windows. Exporting
template instantiations on Windows is an ugly mess for which we have no
machinery whatsoever at present. (At least things should be no more
broken than already...)