COMP: Fill gaps in key explicit instantiation types
This patch fills in explicit instantiation gaps
in the dimensions supported for fixed types.

The list was generated in to provide a consistent
range of coverage of the dimensions needed to support
the ITK toolkit.  Algorithms with fixed dimensions
should be supported upto 10 dimensions.

Using the fixed variants of these matrix/vector
types (rather than dynamically allocarted variants)
provides significant performance improvements.