ENH: Use explict for one argument constructors
Use the C++ keyword explicit for constructors callable with one argument.

Normally, if a constructor can be called with one argument, it can be
used as a conversion. For instance, if you define Foo::Foo(string name)
and then pass a string to a function that expects aFoo, the constructor
will be called to convert the string into a Foo and will pass the Foo to
your function for you. This can be convenient but is also a source of
trouble when things get converted and new objects created without you
meaning them to. Declaring a constructor explicit prevents it from being
invoked implicitly as a conversion.