COMP: Remove unused template export mechanism that is not used
[https://stackoverflow.com/questions/5416872/using-export-keyword-with-templates]
First of all: most compilers (including gcc, Clang and Visual Studio) do
not support the export keyword.

It has been implemented in a single front-end: the EDG front-end, and
thus only the compilers that use it (Comeau and icc) support this
feature. The feedback from the implementers at EDG was extremely simple:
it took us time, was extremely complicated, we recommend not to
implement it (1), as a consequence it has been dropped in C++0x.
                  ---------------------------------------------