BUG: Premature removal of srand48, drand48 restored
NOTE: This patch is a stopgap to restore previous
behavior before the next patch which makes a more
uniform solution across all platforms.

srand48 and drand48 behavior needed to be restored
on unix, and the custom variant for windows needed
to be restored.  While moving to the more standards
conformant C++11 random number generators, those
produce different results than the posix/custom
random number generators.

In order to allow replication of previous
results it is important that the random number seedings
produce the same sequences that were previously
generated.