ENH: Build VXL with WINDOWS_EXPORT_ALL_SYMBOLS & BUILD_SHARED
Windows linkage issues are present with these two options set.

'identifier' : member of dll interface class may not be declared with
dll interface You can declare a whole class, or certain members of a
non-DLL interface class, with DLL interface. You cannot declare a class
with DLL interface and then declare a member of that class with DLL
interface.

'identifier' : definition of dllimport function not allowed Data, static
data members, and functions can be declared as dllimports but not
defined as dllimports.
To fix this issue, remove the __declspec(dllimport) specifier from the definition of the function.