ENH: Consolodate explicit instantiations dims to allsizes
Managing so many 2 line files was tedious and error prone.
Adding explicit instantiations is now much easier, and
maintaining parity between double/float instantiations
is also much easier.  This was motivated by the need
to instantiation more objects in order to get BUILD_SHARED:BOOL=ON
and WINDOWS_EXPORT_ALL_SYMBOLS:BOOL=ON builds to completely
define the required object files for VNL builds.

This PR reduced the number of files to compile by about
100, and greatly reduced the number of header file
reads that are necessary.

Consolidating all dimensions into one file may decrease the
opportunities for parallel compilation, but it also decreases
the disk I/O burden for reading the header files. Given that
most compilations for these explicit instantiations are very
fast, the differential in compile times was not noticably
different.