unify vpgl_backprojet::bproj_plane() interface to accept references
Previously, the generic_camera version accepted the camera as a
reference, and the basic version accepted the camera as a pointer.  With
the new unified interface, overloading can be used to choose the correct
implementation.
Additionally, the vgl interface is converted to a function template so
that it doesn't need to be written for every camera type.