PERF: Add move constructors and move assignment for vnl_vector
The vnl_vector behaviors for move assignment and move
constructors are added.  These need to be carefully
constructed in order to avoid introducing bugs when
the derived vnl_vector_ref class is used.

Careful consideration needs to be made to ensure
that when vnl_[vector/matrix] is the base class
for vnl_[vector/matrix]_ref that the memory is
preserved and respected.

When initializing a vnl_[vector/matrix] object we
ensure that new memory is always created and copied
even when the rvalue is vnl_[vector/matrix] object.

NOTE: currently move-assignment throw an exception, because we need to maintain
      backwards compatibility and the move-construction & move-aasignment
      operators fall back to the copy-assignment operator behavior in
      cases when the memory is externally managed.