COMP: Remove unused netlib function
The adaquad.c file was only used in a test inside
of the netlib v3p folder.  The algorithm was not used
in any other part of vxl.

The adaquad function was identified as having a memory
array bounds overrun bug (from the netlib test) when
included in ITK.

Removing this interface has no impact on VXL, or any
of it's included contrib libraries.