COMP: unary minus result mismatch warning
Resolves #881

The vnl_vector unary minus operator applied to
unsigned type and having a result of unsigned type (it should be signed).

vnl_vector<unsigned short> a(3) { 1, 1, 1 } ;
vnl_vector<short> b = -a;

vnl_vector<short> known= {-1, -1, -1 };

add test to demonstrate incorrect behavior of operator-
When operator-() is used on an unsigned vector type, the results are wrong.