BUG: Fix `is_equal` support when vnl_vector or matrix contains NaN
It appears that `is_equal` would sometimes return `true` when a vnl_vector or
matrix would contain the floating point value NaN. Fixed by adjusting the
condition for doing a `return false` out of the `for` loop, in `is_equal`.

Added unit tests that compare a vnl_vector or matrix that contains NaN and one
that just contains the value zero.