COMP: Replace sprintf to snprintf
Replacement of sprintf to snprintf.

Here only one fix is proposed since it involves doing more involved changes.

Here the snprintf requires the size of buf. Unfortunately, buf, without it’s length, is passed to the function.

We can also see that each function in the file has the same name and a argument list very similar for each of them.

I propose to create at the top a function that has the same name as the others, but with the length of buf added and snprintf used. Then, use that function for the other functions accordingly.

Can I proceed like this for the rest of the functions in the file? Also, should we add a deprecated comment?