COMP: Replace sprintf to snprintf
Replacement of sprintf to snprintf.

Here the snprintf requires the size of buf. Unfortunately, buf, without it’s length, is passed to the function.

We can also see that each function in the file has the same name and a argument list very similar for each of them.

I propose to create a function that has the same name as the others, but with the length of buf added and snprintf used. Then, use that function for the other functions accordingly. Use VXL_DEPRECATED for the function without buf_len. Updated the header file to introduce the newly deprecated functions