BUG: Make speed test more robust to failure
The speed test can fail due when the hardware context switches
during the running of the test.

Put a guard on the speed tests to make this test more stable.  Running
this test on as part of continuous integration environment indicated
that it was not stable.  The CI environment is not well controlled, and
it was possible for the tests to give poor results due to overloaded CI
hardware (or virtualised hardware) making the process of timing very
unstable.  This was observed when the CI systems are under heavy load

This solution adds a guard that at lest 95% of the 20000 tests need
to be faster in the custom 3D computation vs the generic sized version.